Remove NullReferenceException causing code
Remove some code that will never succeed and is guaranteed to cause a
bunch of NullReferenceExceptions when connecting to a server.

They are caught in the Exception catch-all futher down but still show up
in the VS IDE output so stopping them happening in the first place
reduces noise a bit.